removing unused code
