+++ packages/geoview-core/src/core/components/layers/types.ts
@@ -1,56 +1,5 @@
-import {
-  AbstractGeoViewLayer,
-  TypeGeoviewLayerType,
-  TypeLayerEntryConfig,
-  TypeStyleConfig,
-  TypeStyleGeometry,
-  TypeVisibilityFlags,
-} from '@/geo';
+import { TypeGeoviewLayerType, TypeStyleConfig, TypeStyleGeometry, TypeVisibilityFlags } from '@/geo';
 import { TypeLayerStatus } from '@/geo/map/map-schema-types';
-
-export interface TypeLegendProps {
-  layerIds: string[];
-  isRemoveable?: boolean;
-  canSetOpacity?: boolean;
-  expandAll?: boolean;
-  hideAll?: boolean;
-  canZoomTo?: boolean;
-}
-
-export interface LegendItemsProps extends TypeLegendProps {
-  mapId: string;
-}
-
-export interface LegendProps extends TypeLegendProps {
-  mapId: string;
-}
-
-export interface TypeLegendItemProps {
-  layerId: string;
-  geoviewLayerInstance: AbstractGeoViewLayer;
-  subLayerId?: string;
-  layerConfigEntry?: TypeLayerEntryConfig;
-  isRemoveable?: boolean;
-  canSetOpacity?: boolean;
-  isParentVisible?: boolean;
-  toggleParentVisible?: () => void;
-  expandAll?: boolean;
-  hideAll?: boolean;
-}
-
-export interface TypeLegendItemDetailsProps {
-  layerId: string;
-  geoviewLayerInstance: AbstractGeoViewLayer;
-  subLayerId?: string;
-  layerConfigEntry?: TypeLayerEntryConfig;
-  isRemoveable?: boolean;
-  canSetOpacity?: boolean;
-  isParentVisible?: boolean;
-  expandAll?: boolean;
-  hideAll?: boolean;
-}
-
-/// //////////////////////////////////////////////////
 
 export type TypeLayersViewDisplayState = 'remove' | 'add' | 'order' | 'view';
