+++ packages/geoview-core/src/core/components/legend-2/legend.tsx
@@ -2,20 +2,15 @@ import { useStore } from 'zustand';
 import { useTheme } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import { Box, Typography } from '@/ui';
-import { getGeoViewStore } from '@/core/stores/stores-managers';
+import { useGeoViewStore } from '@/core/stores/stores-managers';
 import { getSxClasses } from './legend-styles';
 import { LegendLayer } from './legend-layer';
 
-export interface LegendOverviewProps {
-  mapId: string;
-}
-
-export function Legend(props: LegendOverviewProps): JSX.Element {
-  const { mapId } = props;
+export function Legend(): JSX.Element {
   const theme = useTheme();
   const sxClasses = getSxClasses(theme);
   const { t } = useTranslation<string>();
-  const store = getGeoViewStore(mapId);
+  const store = useGeoViewStore();
   const legendLayers = useStore(store, (state) => state.layerState.legendLayers.filter((f) => f.isVisible));
 
   /* START fake store data here */
